feat: 권한 인증/인가, 초대 기능 구현 (#37)
* docs: Swagger /api Auth

* feat: Req < userId 추가

* feat: [POST] /group manager 설정

* test: /group/create 주석

* feat: `owner`, `editor`, `viewer`

* refactor: max-len 120

* refactor: find > get

* refactor: group.auth < `owner`, `editors`, `viewers`

* feat: pushOwner when create group

* refactor: ParseObjectIdPipe when `groupId` parameter

* refactor: ParseObjectIdPipe when `eventId` parameter

* refactor: ParseObjectIdPipe when `memberId` parameter

* feat: validate group

* feat: group 권한 검증

* feat: 모임 편집자/조회자 초대

* refactor: UserController 정리하고 Swagger 병합

* feat: 모임 중복 초대시 예외처리

* feat: 디스코드 웹훅 ON

* test: rm group.controller.spec.ts

* fix: validate memberIds when deleteMembers

* fix: getEvent > getEvents

* test: TDD는 죽었다

* feat: getEvent

+++ src/group/group.service.ts
@@ -1,18 +1,20 @@
 import { Injectable } from '@nestjs/common';
+import mongoose from 'mongoose';
 
 import { CreateEventDto } from '../event/dto/create-event.dto';
 import { UpdateEventDto } from '../event/dto/update-event.dto';
-import { Event } from '../event/entities/event.entity';
 import { EventService } from '../event/event.service';
 import { CreateMemberDto } from '../member/dto/create-member.dto';
 import { UpdateMemberDto } from '../member/dto/update-member.dto';
 import { MemberService } from '../member/member.service';
 import { GetTransactionsPeriodDto } from '../transaction/dto/get-transaction-period-dto';
 import { TransactionService } from '../transaction/transaction.service';
+import { UserService } from '../user/user.service';
 import { CreateGroupDto } from './dto/create-group.dto';
 import { UpdateGroupDto } from './dto/update-group.dto';
 import { Group } from './entities/group.entity';
 import { GroupRepository } from './group.repository';
+import { GroupValidator } from './group.validator';
 
 @Injectable()
 export class GroupService {
@@ -21,26 +23,31 @@ export class GroupService {
     private readonly memberService: MemberService,
     private readonly eventService: EventService,
     private readonly transactionService: TransactionService,
+    private readonly userService: UserService,
+    private readonly groupValidator: GroupValidator,
   ) {}
 
-  async create(
-    createGroupDto: CreateGroupDto,
-    memberExcel: Express.Multer.File,
-  ): Promise<Group> {
+  async create(userId: string, createGroupDto: CreateGroupDto, memberExcel: Express.Multer.File): Promise<Group> {
     const group: Group = createGroupDto as Group;
+    group.auth = {
+      owner: userId,
+      editors: [],
+      viewers: [],
+    };
 
     if (memberExcel) {
       group.members = await this.memberService.uploadMemberFile(memberExcel);
     }
 
-    return await this.groupRepository.create(group);
+    const createdGroup = await this.groupRepository.create(group);
+    await this.userService.pushOwner(userId, createdGroup.id);
+
+    return createdGroup;
   }
 
-  async addMember(
-    groupId: string,
-    createMemberDto: CreateMemberDto[],
-  ): Promise<Group> {
+  async addMember(userId: string, groupId: string, createMemberDto: CreateMemberDto[]): Promise<Group> {
     const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupEditor(group, userId);
 
     for (const member of createMemberDto) {
       group.members.push(await this.memberService.create(member));
@@ -51,12 +58,10 @@ export class GroupService {
     return group;
   }
 
-  async addMemberToEvent(
-    groupId: string,
-    eventId: string,
-    memberIds: string[],
-  ) {
+  async addMemberToEvent(userId: string, groupId: string, eventId: string, memberIds: string[]) {
     const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupEditor(group, userId);
+
     const event = await this.eventService.getOne(eventId);
 
     for (const member of memberIds) {
@@ -71,11 +76,9 @@ export class GroupService {
     return event;
   }
 
-  async uploadMemberFile(
-    groupId: string,
-    memberExcel: Express.Multer.File,
-  ): Promise<Group> {
-    const group: Group = await this.groupRepository.findOne(groupId);
+  async uploadMemberFile(userId: string, groupId: string, memberExcel: Express.Multer.File): Promise<Group> {
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupEditor(group, userId);
 
     group.members = await this.memberService.uploadMemberFile(memberExcel);
 
@@ -83,22 +86,21 @@ export class GroupService {
   }
 
   async uploadTransactionFile(
+    userId: string,
     groupId: string,
     transactionExcel: Express.Multer.File,
     password: string,
   ): Promise<void> {
-    await this.transactionService.uploadTransactionFile(
-      groupId,
-      transactionExcel,
-      password,
-    );
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupEditor(group, userId);
+
+    await this.transactionService.uploadTransactionFile(groupId, transactionExcel, password);
   }
 
-  async addEvent(
-    groupId: string,
-    createEventDto: CreateEventDto,
-  ): Promise<object> {
+  async addEvent(userId: string, groupId: string, createEventDto: CreateEventDto): Promise<object> {
     const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupEditor(group, userId);
+
     const eventId = await this.eventService.create(createEventDto);
 
     group.events.push(eventId);
@@ -107,12 +109,16 @@ export class GroupService {
     return { eventId };
   }
 
-  async getAll(): Promise<Group[]> {
-    return await this.groupRepository.findAll();
+  async getAll(userId: string): Promise<Group[]> {
+    const user = await this.userService.getOne(userId);
+    const groupIds = user.auth.owner.concat(user.auth.editor).concat(user.auth.viewer);
+
+    return await this.groupRepository.findByIds(groupIds);
   }
 
-  async getOne(groupId: string): Promise<Group> {
+  async getOne(userId: string, groupId: string): Promise<Group> {
     const group: Group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupViewer(group, userId);
 
     const members = await this.getAllMembers(group.members);
     const events = await this.getAllEvents(group.events);
@@ -123,8 +129,10 @@ export class GroupService {
     return group;
   }
 
-  async getMembers(groupId: string) {
+  async getMembers(userId: string, groupId: string) {
     const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupViewer(group, userId);
+
     const members = await this.getAllMembers(group.members);
     const memberInfo = [];
     if (group.members.length > 0) {
@@ -136,51 +144,55 @@ export class GroupService {
     return { memberInfo, members };
   }
 
-  async getEvents(groupId: string) {
-    return this.groupRepository.findOne(groupId).then((group) => {
-      return this.getAllEvents(group.events);
-    });
+  async getEvents(userId: string, groupId: string) {
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupViewer(group, userId);
+
+    return await this.getAllEvents(group.events);
   }
 
-  async getEvent(groupId: string, eventId: string) {
+  async getEvent(userId: string, groupId: string, eventId: string) {
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupViewer(group, userId);
+
     return this.eventService.getOne(eventId);
   }
 
-  async getTransactions(groupId: string) {
+  async getTransactions(userId: string, groupId: string) {
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupViewer(group, userId);
+
     return this.transactionService.getTransactions(groupId);
   }
 
-  async getTransactionsByEvent(groupId: string, eventId: string) {
+  async getTransactionsByEvent(userId: string, groupId: string, eventId: string) {
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupViewer(group, userId);
+
     return this.eventService.compareEventTransactions(groupId, eventId);
   }
 
-  async getTransactionsByPeriod(
-    groupId: string,
-    periodDto: GetTransactionsPeriodDto,
-  ) {
-    return this.transactionService.getTransactionsByPeriod(
-      groupId,
-      periodDto.startDate,
-      periodDto.endDate,
-    );
+  async getTransactionsByPeriod(userId: string, groupId: string, periodDto: GetTransactionsPeriodDto) {
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupViewer(group, userId);
+
+    return this.transactionService.getTransactionsByPeriod(groupId, periodDto.startDate, periodDto.endDate);
   }
 
-  async update(
-    groupId: string,
-    updateGroupDto: UpdateGroupDto,
-  ): Promise<Group> {
+  async update(userId: string, groupId: string, updateGroupDto: UpdateGroupDto): Promise<Group> {
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupEditor(group, userId);
+
     return await this.groupRepository.update(groupId, {
       name: updateGroupDto.name,
       description: updateGroupDto.description,
     } as UpdateGroupDto);
   }
 
-  async updateMember(
-    groupId: string,
-    memberId: string,
-    updateMemberDto: UpdateMemberDto,
-  ) {
+  async updateMember(userId: string, groupId: string, memberId: string, updateMemberDto: UpdateMemberDto) {
     const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupEditor(group, userId);
+
     const member = await this.memberService.update(memberId, updateMemberDto);
 
     group.members = group.members.map((m) => (m === memberId ? member.id : m));
@@ -190,16 +202,11 @@ export class GroupService {
     return member;
   }
 
-  async updateEvent(
-    groupId: string,
-    eventId: string,
-    updateEventDto: UpdateEventDto,
-  ) {
-    const group: Group = await this.groupRepository.findOne(groupId);
-    const event: Event = await this.eventService.update(
-      eventId,
-      updateEventDto,
-    );
+  async updateEvent(userId: string, groupId: string, eventId: string, updateEventDto: UpdateEventDto) {
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupEditor(group, userId);
+
+    const event = await this.eventService.update(eventId, updateEventDto);
 
     group.events = group.events.map((e) => (e === eventId ? event.id : e));
 
@@ -208,30 +215,65 @@ export class GroupService {
     return event;
   }
 
-  async delete(groupId: string): Promise<void> {
+  async delete(userId: string, groupId: string): Promise<void> {
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupOwner(group, userId);
+
     return this.groupRepository.delete(groupId);
   }
 
-  async deleteMembers(groupId: string, memberIds: string[]): Promise<void> {
+  async deleteMembers(userId: string, groupId: string, memberIds: string[]): Promise<void> {
     const group = await this.groupRepository.findOne(groupId);
-    const groupMembers = group.members;
+    this.groupValidator.validateGroupEditor(group, userId);
 
-    group.members = groupMembers.filter(
-      (member) => !memberIds.includes(member),
-    );
+    const groupMembers = group.members;
+    for (const memberId of memberIds) {
+      if (!mongoose.Types.ObjectId.isValid(memberId)) {
+        throw new Error('유효하지 않은 ObjectId입니다.');
+      }
+      if (!groupMembers.includes(memberId)) {
+        throw new Error('멤버가 존재하지 않습니다.');
+      }
+    }
+
+    group.members = groupMembers.filter((member) => !memberIds.includes(member));
     this.memberService.deleteMany(memberIds);
 
     await this.groupRepository.update(groupId, group);
   }
 
-  async deleteEvent(groupId: string, eventId: string): Promise<void> {
+  async deleteEvent(userId: string, groupId: string, eventId: string): Promise<void> {
     const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupEditor(group, userId);
+
     const groupEvents = group.events;
 
     group.events = groupEvents.filter((event) => event !== eventId);
     this.eventService.delete(eventId);
 
     await this.groupRepository.update(groupId, group);
+  }
+
+  async inviteEditor(senderId: string, groupId: string, receiverId: string) {
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupOwner(group, senderId);
+    this.groupValidator.validateGroupInvite(group, receiverId);
+
+    group.auth.editors.push(receiverId);
+    await this.userService.pushEditor(receiverId, groupId);
+
+    return await this.groupRepository.update(groupId, group);
+  }
+
+  async inviteViewer(senderId: string, groupId: string, receiverId: string) {
+    const group = await this.groupRepository.findOne(groupId);
+    this.groupValidator.validateGroupViewer(group, senderId);
+    this.groupValidator.validateGroupInvite(group, receiverId);
+
+    group.auth.viewers.push(receiverId);
+    await this.userService.pushViewer(receiverId, groupId);
+
+    return await this.groupRepository.update(groupId, group);
   }
 
   private async getAllMembers(memberIds: string[]) {
